+++ untitledtvproject/Views/ProfileView.swift
@@ -17,29 +17,29 @@ struct ProfileView: View {
     @State var image: ApiShows.Image?
     
     //language lists
-    @State var englishList : [ApiShows.Returned] = []
-    @State var swedishList : [ApiShows.Returned] = []
-    @State var koreanList : [ApiShows.Returned] = []
-    @State var thaiList : [ApiShows.Returned] = []
-    @State var chineseList : [ApiShows.Returned] = []
-    @State var japaneseList : [ApiShows.Returned] = []
-    @State var danishList : [ApiShows.Returned] = []
-    @State var norwegianList : [ApiShows.Returned] = []
-    @State var germanList : [ApiShows.Returned] = []
-    @State var frenchList : [ApiShows.Returned] = []
-    @State var dutchList : [ApiShows.Returned] = []
-    @State var polishList : [ApiShows.Returned] = []
-    @State var spanishList : [ApiShows.Returned] = []
-    @State var turkishList : [ApiShows.Returned] = []
-    @State var greekList : [ApiShows.Returned] = []
-    @State var allLanguages : [ApiShows.Returned] = []
+    @State var englishList : [ApiShows.ShowReturned] = []
+    @State var swedishList : [ApiShows.ShowReturned] = []
+    @State var koreanList : [ApiShows.ShowReturned] = []
+    @State var thaiList : [ApiShows.ShowReturned] = []
+    @State var chineseList : [ApiShows.ShowReturned] = []
+    @State var japaneseList : [ApiShows.ShowReturned] = []
+    @State var danishList : [ApiShows.ShowReturned] = []
+    @State var norwegianList : [ApiShows.ShowReturned] = []
+    @State var germanList : [ApiShows.ShowReturned] = []
+    @State var frenchList : [ApiShows.ShowReturned] = []
+    @State var dutchList : [ApiShows.ShowReturned] = []
+    @State var polishList : [ApiShows.ShowReturned] = []
+    @State var spanishList : [ApiShows.ShowReturned] = []
+    @State var turkishList : [ApiShows.ShowReturned] = []
+    @State var greekList : [ApiShows.ShowReturned] = []
+    @State var allLanguages : [ApiShows.ShowReturned] = []
     //genre lists
-    @State var dramaList : [ApiShows.Returned] = []
-    @State var comedyList : [ApiShows.Returned] = []
-    @State var scifiList : [ApiShows.Returned] = []
-    @State var horrorList : [ApiShows.Returned] = []
-    @State var crimeList : [ApiShows.Returned] = []
-    @State var adventureList : [ApiShows.Returned] = []
+    @State var dramaList : [ApiShows.ShowReturned] = []
+    @State var comedyList : [ApiShows.ShowReturned] = []
+    @State var scifiList : [ApiShows.ShowReturned] = []
+    @State var horrorList : [ApiShows.ShowReturned] = []
+    @State var crimeList : [ApiShows.ShowReturned] = []
+    @State var adventureList : [ApiShows.ShowReturned] = []
         
     var user = Auth.auth().currentUser
     let storage = Storage.storage() //storage for user selected profile picture
@@ -361,7 +361,7 @@ struct ProfileView: View {
                 clearLists()
                 for document in snapshot.documents {
                     let result = Result {
-                        try document.data(as: ApiShows.Returned.self)
+                        try document.data(as: ApiShows.ShowReturned.self)
                     }
                     switch result  {
                     case .success(let show)  :
