+++ untitledtvproject/Views/ShowEntryView.swift
@@ -15,7 +15,7 @@ import FirebaseFirestore
 
 struct ShowEntryView : View {
     
-    @State var show2 : ApiShows.Returned
+    @State var show2 : ApiShows.ShowReturned
     @State var name : String = ""
     @State var language : String = ""
     @State var summary: String = ""
@@ -46,7 +46,7 @@ struct ShowEntryView : View {
                 .padding(.top, -10)
                 .padding(10)
                 NavigationLink {
-                    ShowEntryMoreView(show2: show2, name: show2.show.name, language: show2.show.language, summary: show2.show.summary, image: show2.show.image, type: show2.show.type, network: show2.show.network, status: show2.show.status, premiered: show2.show.premiered, rating: show2.show.rating!)
+                    EntryDetailsView(show2: show2, name: show2.show.name, language: show2.show.language, summary: show2.show.summary, image: show2.show.image, type: show2.show.type, network: show2.show.network, status: show2.show.status, premiered: show2.show.premiered, rating: show2.show.rating!)
                 } label: {
                     Text("Show more information")
                 }
